renames package and natie-image.properties path
