+++ micrometer-registry-new-relic-telemetry/src/main/java/io/micronaut/configuration/metrics/micrometer/newrelictelemetry/package-info.java
@@ -21,7 +21,7 @@
  */
 @Configuration
 @Requires(property = NewRelicTelemetryMeterRegistryFactory.NEWRELIC_ENABLED, notEquals = StringUtils.FALSE)
-package io.micronaut.configuration.metrics.micrometer.newrelic;
+package io.micronaut.configuration.metrics.micrometer.newrelictelemetry;
 
 import io.micronaut.context.annotation.Configuration;
 import io.micronaut.context.annotation.Requires;
