+++ micrometer-registry-new-relic-telemetry/src/main/resources/META-INF/native-image/io.micronaut.configuration.metrics/micrometer-new-relic-telemetry/native-image.properties
